+++ tanaguru-engine/tanaguru-audit-runner/src/main/java/com/tanaguru/runner/factory/AuditRunnerFactoryImpl.java
@@ -18,6 +18,15 @@ import com.tanaguru.repository.ScenarioRepository;
 import com.tanaguru.repository.TanaguruTestRepository;
 import com.tanaguru.runner.*;
 import com.tanaguru.service.AuditService;
+import org.apache.http.HttpHost;
+import org.apache.http.auth.AuthScope;
+import org.apache.http.auth.UsernamePasswordCredentials;
+import org.apache.http.client.CredentialsProvider;
+import org.apache.http.client.HttpClient;
+import org.apache.http.conn.params.ConnRoutePNames;
+import org.apache.http.impl.client.BasicCredentialsProvider;
+import org.apache.http.impl.client.DefaultHttpClient;
+import org.apache.http.impl.client.HttpClientBuilder;
 import org.openqa.selenium.remote.RemoteWebDriver;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -25,6 +34,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 import javax.transaction.Transactional;
+import java.net.ProxySelector;
 import java.util.*;
 import java.util.stream.Collectors;