+++ CHANGELOG.md
@@ -28,3 +28,4 @@
 - Optimize TestResult getAllTestResultByReference endpoint
 - Add a parameter on createUser endpoint to create a contract
 - Upgrade logging
+- Fix audit and user deletion